# goldtag-api/swaggerDoc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Goldtag App",
    "contact": {
      "name": "Goldtag App",
      "url": "https://goldtag.org",
      "email": "info@goldtag.net"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "variables": {}
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "register user",
        "description": "Registers user",
        "operationId": "registeruser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/registeruserrequest"
                  },
                  {
                    "example": {
                      "name": "peterpan",
                      "email": "peter3@gmail.com",
                      "password": "LKkBUIblm-o0"
                    }
                  }
                ]
              },
              "example": {
                "name": "peterpan",
                "email": "peter3@gmail.com",
                "password": "LKkBUIblm-o0"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Register Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/successresponse"
                    },
                    {
                      "example": {
                        "success": true,
                        "message": "Signing successfull ",
                        "data": {
                          "user": {
                            "name": "peterpan",
                            "email": "peter3@gmail.com"
                          },
                          "token": "9jkk5nmterjbtokp453hujrh34hugq3rj43oi29yr39ofhoiwe8ehf8934hggf98723g8wy47tgf7w3gu9wg7rf8wehf9wf"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/badrequest"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userconflict"
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "login user",
        "description": "Login as user",
        "operationId": "loginuser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/loginuserrequest"
                  },
                  {
                    "example": {
                      "email": "peter3@gmail.com",
                      "password": "LKkBUIblm-o0"
                    }
                  }
                ]
              },
              "example": {
                "email": "peter3@gmail.com",
                "password": "LKkBUIblm-o0"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/successresponse"
                    },
                    {
                      "example": {
                        "success": true,
                        "message": "Signing successfull ",
                        "data": {
                          "user": {
                            "name": "peterpan",
                            "email": "peter3@gmail.com"
                          },
                          "token": "9jkk5nmterjbtokp453hujrh34hugq3rj43oi29yr39ofhoiwe8ehf8934hggf98723g8wy47tgf7w3gu9wg7rf8wehf9wf"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "No user found with that email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/badrequest"
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "registeruserrequest": {
        "title": "registeruserrequest",
        "required": ["name", "email", "password"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "name": "peterpan",
          "email": "peter3@gmail.com",
          "password": "LKkBUIblm-o0"
        }
      },
      "loginuserrequest": {
        "title": "loginuserrequest",
        "required": ["email", "password"],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "peter3@gmail.com",
          "password": "LKkBUIblm-o0"
        }
      },
...
      "successresponse": {
        "title": "success",
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "badrequest": {
        "title": "badrequest",
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          }
        },
        "example": {
          "success": false,
          "error": "Bad request please type in valid values/parameters"
        }
      },
      "userconflict": {
        "title": "conflict",
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          }
        },
        "example": {
          "success": false,
          "error": "User already exist in DB with given email please type in another value"
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}